add: Included error handling and message display add: Add authentication for creating a new Subscription add: Added required APIs in Subscription and Subscribe fix: Generalised api responses fix: Fixed bug in publisher authentication

+++ As-5/subscription/src/main/java/com/newsletter/subscription/controller/SubscriptionController.java
@@ -1,8 +1,13 @@
 package com.newsletter.subscription.controller;
 
-import com.newsletter.subscription.constants.Constant;
-import com.newsletter.subscription.dto.SubscriptionDto;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.newsletter.subscription.constants.MessageConstant;
+import com.newsletter.subscription.constants.PathConstant;
+import com.newsletter.subscription.constants.UtilityConstant;
+import com.newsletter.subscription.dto.*;
+import com.newsletter.subscription.feignclient.UserServiceClient;
 import com.newsletter.subscription.mapper.SubscriptionMapper;
+import com.newsletter.subscription.models.Subscription;
 import com.newsletter.subscription.services.SubscriptionService;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +15,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 /**
@@ -17,12 +23,17 @@ import java.util.stream.Collectors;
  */
 @AllArgsConstructor
 @RestController
-@RequestMapping(Constant.SUBSCRIPTIONS_MAPPING)
+@RequestMapping(PathConstant.SUBSCRIPTIONS_MAPPING)
 public class SubscriptionController {
 
   @Autowired
   private final SubscriptionService subscriptionService;
+  @Autowired
   private final SubscriptionMapper subscriptionMapper;
+  @Autowired
+  private final UserServiceClient userServiceClient;
+
+  private final ObjectMapper objectMapper;
 
   /**
    * Create response entity.
@@ -31,30 +42,95 @@ public class SubscriptionController {
    * @return the response entity
    */
   @PostMapping
-  ResponseEntity<SubscriptionDto> create(@RequestBody SubscriptionDto subscriptionDto) {
-    return ResponseEntity.ok(
-        subscriptionMapper.subscriptionToDto(
-            subscriptionService.create(
-                subscriptionMapper.dtoToSubscription(subscriptionDto)
-            )
-        )
-    );
+  ResponseEntity<ApiResponseDto> create(
+      @RequestBody SubscriptionDto subscriptionDto,
+      @RequestHeader(UtilityConstant.SECURITY_HEADER) String bearer) {
+
+    ApiResponseDto apiResponseDto = new ApiResponseDto();
+
+    try {
+
+      // Authorize publisher
+      boolean authorized = Objects.requireNonNull(userServiceClient
+              .authorizePublisher(bearer)
+              .getBody())
+          .getStatus()
+          .equalsIgnoreCase(MessageConstant.ACCEPTED);
+
+      if (authorized) {
+        Subscription subscription = subscriptionMapper
+            .dtoToSubscription(subscriptionDto);
+
+        // Get publisher
+        UserResponseDto userResponseDto =
+            objectMapper.convertValue(
+                Objects.requireNonNull(userServiceClient
+                        .getUserByToken(bearer)
+                        .getBody())
+                    .getBody(),
+                UserResponseDto.class
+            );
+        // Add publisher id to subscription
+
+        subscription.setPublisherId(userResponseDto.getId());
+
+        // Create Subscription
+        subscriptionDto =
+            subscriptionMapper
+                .subscriptionToDto(subscriptionService
+                    .create(subscription));
+        apiResponseDto.setBody(subscriptionDto);
+
+        apiResponseDto.setStatus(MessageConstant.ACCEPTED);
+        apiResponseDto.setMessage(MessageConstant.SUCCESS);
+
+      } else {
+        apiResponseDto.setStatus(MessageConstant.FAILED);
+        apiResponseDto.setMessage(MessageConstant.UNAUTHORIZED);
+      }
+    } catch (Exception exception) {
+      exception.printStackTrace();
+
+      apiResponseDto.setStatus(MessageConstant.DENIED);
+      apiResponseDto.setMessage(MessageConstant.SUBSCRIPTION_CREATION_FAILED);
+    }
+
+    return ResponseEntity.ok(apiResponseDto);
+
   }
 
   /**
-   * Gets all.
+   * Gets all subscriptions.
    *
-   * @return the all
+   * @return the list of subscriptions
    */
   @GetMapping
-  ResponseEntity<List<SubscriptionDto>> getAll() {
-    return ResponseEntity.ok(
-        subscriptionService
-            .getAll()
-            .stream()
-            .map(subscriptionMapper::subscriptionToDto)
-            .collect(Collectors.toList())
-    );
+  ResponseEntity<ApiResponseDto> getAll() {
+
+
+    ApiResponseDto apiResponseDto = new ApiResponseDto();
+
+    try {
+
+      List<SubscriptionDto> subscriptionDtoList = subscriptionService
+          .getAll()
+          .stream()
+          .map(subscriptionMapper::subscriptionToDto)
+          .collect(Collectors.toList());
+
+      apiResponseDto.setBody(subscriptionDtoList);
+      apiResponseDto.setStatus(MessageConstant.ACCEPTED);
+      apiResponseDto.setMessage(MessageConstant.SUCCESS);
+
+    } catch (Exception exception) {
+      exception.printStackTrace();
+
+      apiResponseDto.setStatus(MessageConstant.DENIED);
+      apiResponseDto.setMessage(MessageConstant.SUBSCRIPTION_FETCH_FAILED);
+    }
+
+    return ResponseEntity.ok(apiResponseDto);
+
   }
 
   /**
@@ -63,13 +139,27 @@ public class SubscriptionController {
    * @param id the id
    * @return the by id
    */
-  @GetMapping(Constant.BY_ID_MAPPING)
-  ResponseEntity<SubscriptionDto> getById(@PathVariable String id) {
-    return ResponseEntity.ok(
-        subscriptionMapper.subscriptionToDto(
-            subscriptionService.getById(id)
-        )
-    );
+  @GetMapping(PathConstant.BY_ID_MAPPING)
+  ResponseEntity<ApiResponseDto> getById(@PathVariable String id) {
+
+    ApiResponseDto apiResponseDto = new ApiResponseDto();
+
+    try {
+
+      SubscriptionDto subscriptionDto = subscriptionMapper
+          .subscriptionToDto(subscriptionService.getById(id));
+
+      apiResponseDto.setBody(subscriptionDto);
+      apiResponseDto.setStatus(MessageConstant.ACCEPTED);
+      apiResponseDto.setMessage(MessageConstant.SUCCESS);
+    } catch (Exception exception) {
+      exception.printStackTrace();
+
+      apiResponseDto.setStatus(MessageConstant.DENIED);
+      apiResponseDto.setMessage(MessageConstant.SUBSCRIPTION_NOT_FOUND);
+    }
+
+    return ResponseEntity.ok(apiResponseDto);
   }
 
   /**
@@ -78,29 +168,132 @@ public class SubscriptionController {
    * @param query the query
    * @return the response entity
    */
-  @GetMapping(Constant.SEARCH_MAPPING)
-  ResponseEntity<List<SubscriptionDto>> search (@RequestParam String query) {
-    return ResponseEntity.ok(subscriptionService.search(query).stream().map(
-        subscriptionMapper::subscriptionToDto
-    ).collect(Collectors.toList()));
+  @GetMapping(PathConstant.SEARCH_MAPPING + PathConstant.BY_NAME_MAPPING)
+  ResponseEntity<ApiResponseDto> searchByName (@RequestParam String query) {
+
+    ApiResponseDto apiResponseDto = new ApiResponseDto();
+
+    try {
+
+      List<SubscriptionDto> subscriptionDtoList = subscriptionService
+          .searchByName(query)
+          .stream()
+          .map(subscriptionMapper::subscriptionToDto)
+          .collect(Collectors.toList());
+
+      apiResponseDto.setBody(subscriptionDtoList);
+      apiResponseDto.setStatus(MessageConstant.ACCEPTED);
+      apiResponseDto.setMessage(MessageConstant.SUCCESS);
+    } catch (Exception exception) {
+      exception.printStackTrace();
+
+      apiResponseDto.setStatus(MessageConstant.DENIED);
+      apiResponseDto.setMessage(MessageConstant.SUBSCRIPTION_FETCH_FAILED);
+    }
+
+    return ResponseEntity.ok(apiResponseDto);
   }
 
-  /**
-   * Update response entity.
-   *
-   * @param subscriptionDto the subscription dto
-   * @param id              the id
-   * @return the response entity
-   */
-  @PutMapping(Constant.BY_ID_MAPPING)
-  ResponseEntity<SubscriptionDto> update(@RequestBody SubscriptionDto subscriptionDto, @PathVariable String id) {
-    return ResponseEntity.ok(
-        subscriptionMapper.subscriptionToDto(
+  @GetMapping(PathConstant.SEARCH_MAPPING + PathConstant.BY_CATEGORY_MAPPING)
+  ResponseEntity<ApiResponseDto> searchByCategory (@RequestParam String query) {
+
+    ApiResponseDto apiResponseDto = new ApiResponseDto();
+
+    try {
+
+      List<SubscriptionDto> subscriptionDtoList = subscriptionService
+          .searchByCategory(query)
+          .stream()
+          .map(subscriptionMapper::subscriptionToDto)
+          .collect(Collectors.toList());
+
+      apiResponseDto.setBody(subscriptionDtoList);
+      apiResponseDto.setStatus(MessageConstant.ACCEPTED);
+      apiResponseDto.setMessage(MessageConstant.SUCCESS);
+    } catch (Exception exception) {
+      exception.printStackTrace();
+
+      apiResponseDto.setStatus(MessageConstant.DENIED);
+      apiResponseDto.setMessage(MessageConstant.SUBSCRIPTION_FETCH_FAILED);
+    }
+
+    return ResponseEntity.ok(apiResponseDto);
+  }
+
+  @GetMapping(PathConstant.SEARCH_MAPPING + PathConstant.BY_PUBLISHER_ID)
+  ResponseEntity<ApiResponseDto> searchByPublisherId (@RequestParam Long query) {
+
+    ApiResponseDto apiResponseDto = new ApiResponseDto();
+
+    try {
+
+      List<SubscriptionDto> subscriptionDtoList = subscriptionService
+          .searchByPublisherId(query)
+          .stream()
+          .map(subscriptionMapper::subscriptionToDto)
+          .collect(Collectors.toList());
+
+      apiResponseDto.setBody(subscriptionDtoList);
+      apiResponseDto.setStatus(MessageConstant.ACCEPTED);
+      apiResponseDto.setMessage(MessageConstant.SUCCESS);
+    } catch (Exception exception) {
+      exception.printStackTrace();
+
+      apiResponseDto.setStatus(MessageConstant.DENIED);
+      apiResponseDto.setMessage(MessageConstant.SUBSCRIPTION_FETCH_FAILED);
+    }
+
+    return ResponseEntity.ok(apiResponseDto);
+  }
+
+  @PutMapping(PathConstant.BY_ID_MAPPING)
+  ResponseEntity<ApiResponseDto> update(
+      @RequestBody SubscriptionDto subscriptionDto,
+      @PathVariable String id,
+      @RequestHeader(UtilityConstant.SECURITY_HEADER) String bearer) {
+
+    ApiResponseDto apiResponseDto = new ApiResponseDto();
+
+    try {
+
+      // Authorize publisher(owner of subscription)
+      boolean authorized = Objects.requireNonNull(userServiceClient
+              .authorizePublisher(bearer)
+              .getBody())
+          .getStatus()
+          .equalsIgnoreCase(MessageConstant.ACCEPTED);
+
+      // Get publisher
+      UserResponseDto userResponseDto =
+          objectMapper.convertValue(
+              Objects.requireNonNull(userServiceClient
+                      .getUserByToken(bearer)
+                      .getBody())
+                  .getBody(),
+              UserResponseDto.class
+          );
+      if (authorized && Objects.equals(userResponseDto.getId(), subscriptionDto.getPublisherId())) {
+
+        // Update subscription
+        subscriptionDto = subscriptionMapper.subscriptionToDto(
             subscriptionService.update(
-                subscriptionMapper.dtoToSubscription(subscriptionDto), id
-            )
-        )
-    );
+                subscriptionMapper.dtoToSubscription(subscriptionDto), id));
+        apiResponseDto.setBody(subscriptionDto);
+
+        apiResponseDto.setStatus(MessageConstant.ACCEPTED);
+        apiResponseDto.setMessage(MessageConstant.SUCCESS);
+      } else {
+        apiResponseDto.setStatus(MessageConstant.FAILED);
+        apiResponseDto.setMessage(MessageConstant.UNAUTHORIZED);
+      }
+    } catch (Exception exception) {
+      exception.printStackTrace();
+
+      apiResponseDto.setStatus(MessageConstant.DENIED);
+      apiResponseDto.setMessage(MessageConstant.SUBSCRIPTION_UPDATE_FAILED);
+    }
+
+    return ResponseEntity.ok(apiResponseDto);
   }
 
   /**
@@ -108,9 +301,50 @@ public class SubscriptionController {
    *
    * @param id the id
    */
-  @DeleteMapping(Constant.BY_ID_MAPPING)
-  void delete(@PathVariable String id) {
-    subscriptionService.delete(id);
+  @DeleteMapping(PathConstant.BY_ID_MAPPING)
+  ResponseEntity<ApiResponseDto> delete(
+      @PathVariable String id,
+      @RequestHeader(UtilityConstant.SECURITY_HEADER) String bearer) {
+
+    ApiResponseDto apiResponseDto = new ApiResponseDto();
+
+    try {
+      // Authorize publisher(owner of subscription)
+      boolean authorized = Objects.requireNonNull(userServiceClient
+              .authorizePublisher(bearer)
+              .getBody())
+          .getStatus()
+          .equalsIgnoreCase(MessageConstant.ACCEPTED);
+
+      // Get publisher
+      UserResponseDto userResponseDto =
+          objectMapper.convertValue(
+              Objects.requireNonNull(userServiceClient
+                      .getUserByToken(bearer)
+                      .getBody())
+                  .getBody(),
+              UserResponseDto.class
+          );
+      if (authorized &&
+          Objects.equals(userResponseDto.getId(), subscriptionService.getById(id).getPublisherId())) {
+
+        // Delete subscription
+        subscriptionService.delete(id);
+
+        apiResponseDto.setStatus(MessageConstant.ACCEPTED);
+        apiResponseDto.setMessage(MessageConstant.SUCCESS);
+      } else {
+        apiResponseDto.setStatus(MessageConstant.FAILED);
+        apiResponseDto.setMessage(MessageConstant.UNAUTHORIZED);
+      }
+    } catch (Exception exception) {
+      exception.printStackTrace();
+
+      apiResponseDto.setStatus(MessageConstant.DENIED);
+      apiResponseDto.setMessage(MessageConstant.SUBSCRIPTION_DELETION_FAILED);
+    }
+
+    return ResponseEntity.ok(apiResponseDto);
   }
 
 }
